requirements.txt
flask==3.0.2
google-generativeai==0.3.2
pydantic==2.6.3
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.2.0
flask-cors==4.0.0 